+++ src/main/java/fr/esgi/cookRecipe/Exposition/SocialDTO/RatingUserDTO.java
@@ -6,11 +6,11 @@ public class RatingUserDTO {
     private boolean userRated;
     private double userRating;
 
-    public static RatingUserDTO of(RatingDTO rating, boolean userRated, double userRating) {
+    public static RatingUserDTO of(RatingDTO rating, boolean userRated, int userRating) {
         return new RatingUserDTO(rating, userRated, userRating);
     }
 
-    private RatingUserDTO(RatingDTO rating, boolean userRated, double userRating) {
+    private RatingUserDTO(RatingDTO rating, boolean userRated, int userRating) {
         this.rating = rating;
         this.userRated = userRated;
         this.userRating = userRating;
